requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.2.0
black==20.8b1
click==7.1.2
colorama==0.4.3
ExifRead==2.3.1
ffmpeg-python==0.2.0
future==0.18.2
importlib-metadata==2.0.0
iniconfig==1.0.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.0
regex==2020.9.27
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.2.0
